Move {query,get}SnsNeuron{,s} from sns.api to sns-governance.api
