+++ frontend/src/lib/services/sns-neurons-check-balances.services.ts
@@ -3,7 +3,7 @@ import {
   getNeuronBalance,
   refreshNeuron,
 } from "$lib/api/sns-governance.api";
-import { getSnsNeuron } from "$lib/api/sns.api";
+import { getSnsNeuron } from "$lib/api/sns-governance.api";
 import { MAX_NEURONS_SUBACCOUNTS } from "$lib/constants/sns-neurons.constants";
 import { getAuthenticatedIdentity } from "$lib/services/auth.services";
 import { snsNeuronsStore } from "$lib/stores/sns-neurons.store";
